requirements-test.txt
pytest
pytest-cov
pytest-asyncio
pytest-homeassistant-custom-component